% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aodn.r
\name{read_aodn_csv}
\alias{read_aodn_csv}
\title{Read CSV from the AODN portal.}
\usage{
read_aodn_csv(u, ...)
}
\arguments{
\item{u}{AODN CSV URL}

\item{...}{arguments passed to methods}
}
\value{
raodn data frame
}
\description{
Returns a data.frame as is with a simple S3 class for future use.
}
\examples{

d <- read_aodn_csv(collection_url)
#library(ggplot2)
# ggplot(d[d$pres < 100, ]) + aes(longitude, latitude, col = psal) + geom_point()
#ggplot(d) + aes(psal, -pres, col = temp) +
# geom_point() +
#  facet_wrap(~platform_number)
}
